third_party/WebKit/LayoutTests/platform/mac-mavericks/fast/text/custom-font-data-crash2-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: custom-font-data-crash2.html
CONSOLE WARNING: OTS parsing error: invalid version tag
CONSOLE WARNING: Failed to decode downloaded font: custom-font-data-crash2.html
CONSOLE WARNING: OTS parsing error: invalid version tag
PASS
